requirements.txt
matplotlib>=3.8
numpy>=1.24
opencv-python-headless>=4.10
pandas>=2.2
pillow>=10.4
grad-cam @ git+https://github.com/jacobgil/pytorch-grad-cam.git
scikit-learn>=1.5
timm>=1.0
torch>=2.0
torchvision>=0.15
